# backend/docs/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Housing Society Management API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/properties/locations": {
      "get": {
        "tags": [
          "properties"
        ],
        "summary": "List Locations",
        "operationId": "list_locations_api_v1_properties_locations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SocietyLocationOut"
                  },
                  "type": "array",
                  "title": "Response List Locations Api V1 Properties Locations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "properties"
        ],
        "summary": "Create Location",
        "operationId": "create_location_api_v1_properties_locations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocietyLocationCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocietyLocationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/properties": {
      "get": {
        "tags": [
          "properties"
        ],
        "summary": "List Properties",
        "operationId": "list_properties_api_v1_properties_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropertyOut"
                  },
                  "type": "array",
                  "title": "Response List Properties Api V1 Properties Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "properties"
        ],
        "summary": "Create Property",
        "operationId": "create_property_api_v1_properties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/residents/{resident_id}/properties": {
      "get": {
        "tags": [
          "residents"
        ],
        "summary": "Get Resident Properties",
        "operationId": "get_resident_properties_api_v1_residents__resident_id__properties_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Resident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyResidentOut"
                  },
                  "title": "Response Get Resident Properties Api V1 Residents  Resident Id  Properties Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/proposals": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_proposals_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalCreateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "proposals"
        ],
        "summary": "List All",
        "operationId": "list_all_api_v1_proposals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProposalOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{proposal_id}": {
      "get": {
        "tags": [
          "proposals"
        ],
        "summary": "Get Detail",
        "operationId": "get_detail_api_v1_proposals__proposal_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalStatusDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{proposal_id}/vote": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Vote",
        "operationId": "vote_api_v1_proposals__proposal_id__vote_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CastVoteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/{proposal_id}/withdraw": {
      "post": {
        "tags": [
          "proposals"
        ],
        "summary": "Withdraw",
        "operationId": "withdraw_api_v1_proposals__proposal_id__withdraw_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CastVoteIn": {
        "properties": {
          "vote": {
            "$ref": "#/components/schemas/Vote"
          }
        },
        "type": "object",
        "required": [
          "vote"
        ],
        "title": "CastVoteIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HouseType": {
        "type": "string",
        "enum": [
          "FLAT",
          "BUNGALOW"
        ],
        "title": "HouseType"
      },
      "LocationType": {
        "type": "string",
        "enum": [
          "WING",
          "ROW"
        ],
        "title": "LocationType"
      },
...
      "Page_ProposalOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProposalOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[ProposalOut]"
      },
...
      "PropertyCreateIn": {
        "properties": {
          "location_id": {
            "type": "string",
            "format": "uuid",
            "title": "Location Id"
          },
          "house_number": {
            "type": "string",
            "title": "House Number"
          },
          "house_type": {
            "$ref": "#/components/schemas/HouseType"
          },
          "floor_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor Number"
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "house_number",
          "house_type"
        ],
        "title": "PropertyCreateIn"
      },
      "PropertyOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "location_id": {
            "type": "string",
            "format": "uuid",
            "title": "Location Id"
          },
          "house_number": {
            "type": "string",
            "title": "House Number"
          },
          "house_type": {
            "$ref": "#/components/schemas/HouseType"
          },
          "floor_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "location_id",
          "house_number",
          "house_type",
          "floor_number",
          "status",
          "created_at"
        ],
        "title": "PropertyOut"
      },
...
      "PropertyResidentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "relationship_type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "property_id",
          "resident_id",
          "relationship_type",
          "is_active",
          "start_date",
          "end_date"
        ],
        "title": "PropertyResidentOut"
      },
      "ProposalCreateIn": {
        "properties": {
          "scope_type": {
            "$ref": "#/components/schemas/ProposalScope"
          },
          "scope_location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Location Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "scope_type",
          "title",
          "description"
        ],
        "title": "ProposalCreateIn"
      },
      "ProposalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "scope_type": {
            "$ref": "#/components/schemas/ProposalScope"
          },
          "scope_location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Location Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "scope_type",
          "scope_location_id",
          "title",
          "description",
          "status",
          "created_by",
          "withdrawn_at",
          "created_at"
        ],
        "title": "ProposalOut"
      },
      "ProposalScope": {
        "type": "string",
        "enum": [
          "SOCIETY",
          "WING",
          "ROW"
        ],
        "title": "ProposalScope"
      },
      "ProposalStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "APPROVED",
          "WITHDRAWN"
        ],
        "title": "ProposalStatus"
      },
      "ProposalStatusDetailOut": {
        "properties": {
          "proposal": {
            "$ref": "#/components/schemas/ProposalOut"
          },
          "resident_approve_count": {
            "type": "integer",
            "title": "Resident Approve Count"
          },
          "resident_eligible_count": {
            "type": "integer",
            "title": "Resident Eligible Count"
          },
          "resident_percent": {
            "type": "number",
            "title": "Resident Percent"
          },
          "subadmin_approve_count": {
            "type": "integer",
            "title": "Subadmin Approve Count"
          },
          "subadmin_eligible_count": {
            "type": "integer",
            "title": "Subadmin Eligible Count"
          },
          "subadmin_percent": {
            "type": "number",
            "title": "Subadmin Percent"
          },
          "resident_threshold_met": {
            "type": "boolean",
            "title": "Resident Threshold Met"
          },
          "subadmin_threshold_met": {
            "type": "boolean",
            "title": "Subadmin Threshold Met"
          }
        },
        "type": "object",
        "required": [
          "proposal",
          "resident_approve_count",
          "resident_eligible_count",
          "resident_percent",
          "subadmin_approve_count",
          "subadmin_eligible_count",
          "subadmin_percent",
          "resident_threshold_met",
          "subadmin_threshold_met"
        ],
        "title": "ProposalStatusDetailOut",
        "description": "Live threshold snapshot at read time \u2014 never cached (Section 49.1)."
      },
...
      "RelationshipType": {
        "type": "string",
        "enum": [
          "OWNER",
          "TENANT"
        ],
        "title": "RelationshipType"
      },
...
      "SocietyLocationCreateIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location_type": {
            "$ref": "#/components/schemas/LocationType"
          }
        },
        "type": "object",
        "required": [
          "name",
          "location_type"
        ],
        "title": "SocietyLocationCreateIn"
      },
      "SocietyLocationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location_type": {
            "$ref": "#/components/schemas/LocationType"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "name",
          "location_type",
          "created_at"
        ],
        "title": "SocietyLocationOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Vote": {
        "type": "string",
        "enum": [
          "APPROVE",
          "REJECT"
        ],
        "title": "Vote"
      },
      "VoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "proposal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Proposal Id"
          },
          "voter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Voter Id"
          },
          "vote": {
            "$ref": "#/components/schemas/Vote"
          },
          "voted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Voted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "proposal_id",
          "voter_id",
          "vote",
          "voted_at"
        ],
        "title": "VoteOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}